Move asseertable interfaces
