--- playtest-core/src/main/kotlin/com/uzabase/playtest2/core/assertion/Assertable.kt
@@ -1,26 +1,0 @@
-package com.uzabase.playtest2.core.assertion
-
-fun interface ShouldBeLong {
-    fun shouldBe(expected: Long): Boolean
-}
-
-fun interface ShouldBeGreaterEqualLong {
-    fun shouldBeGreaterEqual(expected: Long): Boolean
-}
-
-fun interface ShouldBeString {
-    fun shouldBe(expected: String): Boolean
-}
-
-fun interface ShouldContainsString {
-    fun shouldContain(expected: String): Boolean
-}
-
-fun interface ShouldBeBoolean {
-    fun shouldBe(expected: Boolean): Boolean
-}
-
-fun interface ShouldBe<T> {
-    fun shouldBe(expected: T): Boolean
-}
-
