+++ playtest-core/src/main/kotlin/com/uzabase/playtest2/core/assertion/ForBoolean.kt
@@ -1,0 +1,5 @@
+package com.uzabase.playtest2.core.assertion
+
+fun interface ShouldBeBoolean {
+    fun shouldBe(expected: Boolean): Boolean
+}
