+++ playtest-core/src/main/kotlin/com/uzabase/playtest2/core/assertion/ForNumber.kt
@@ -1,0 +1,9 @@
+package com.uzabase.playtest2.core.assertion
+
+fun interface ShouldBeLong {
+    fun shouldBe(expected: Long): Boolean
+}
+
+fun interface ShouldBeGreaterEqualLong {
+    fun shouldBeGreaterEqual(expected: Long): Boolean
+}
