+++ playtest-core/src/main/kotlin/com/uzabase/playtest2/core/assertion/ForString.kt
@@ -1,0 +1,9 @@
+package com.uzabase.playtest2.core.assertion
+
+fun interface ShouldBeString {
+    fun shouldBe(expected: String): Boolean
+}
+
+fun interface ShouldContainsString {
+    fun shouldContain(expected: String): Boolean
+}
