+++ playtest-core/src/main/kotlin/com/uzabase/playtest2/core/assertion/ForT.kt
@@ -1,0 +1,5 @@
+package com.uzabase.playtest2.core.assertion
+
+fun interface ShouldBe<T> {
+    fun shouldBe(expected: T): Boolean
+}
