+++ playtest-core/src/main/kotlin/com/uzabase/playtest2/core/assertion/PlaytestAssertionError.kt
@@ -1,0 +1,5 @@
+package com.uzabase.playtest2.core.assertion
+
+class PlaytestAssertionError(override val message: String) : AssertionError(message)
+
+fun playtestError(message: String): Nothing = throw PlaytestAssertionError(message)
