--- playtest-core/src/main/kotlin/com/uzabase/playtest2/core/assertion/PlaytestException.kt
@@ -1,5 +1,0 @@
-package com.uzabase.playtest2.core.assertion
-
-class PlaytestException(override val message: String) : AssertionError(message)
-
-fun playtestException(message: String): Nothing = throw PlaytestException(message)
