+++ playtest-core/src/test/kotlin/com/uzabase/playtest2/core/ScenarioDataStoreExtTest.kt
@@ -1,7 +1,7 @@
 package com.uzabase.playtest2.core
 
 import com.thoughtworks.gauge.datastore.ScenarioDataStore
-import com.uzabase.playtest2.core.assertion.PlaytestException
+import com.uzabase.playtest2.core.assertion.PlaytestAssertionError
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.nulls.shouldBeNull
@@ -24,7 +24,7 @@ class ScenarioDataStoreExtTest : FunSpec({
 
         test("should throw PlaytestException if the value is not the expected type") {
             ScenarioDataStore.put("doge", Doge("space doge"))
-            shouldThrow<PlaytestException> {
+            shouldThrow<PlaytestAssertionError> {
                 ScenarioDataStoreExt.getAs<List<Doge>>("doge")
             }.message.shouldStartWith("ScenarioDataStore does not contain key `doge` as `List`.")
         }
